create separate page for adventure
