+++ index.html
@@ -20,7 +20,10 @@
       href="https://fonts.googleapis.com/css?family=Montserrat|Open+Sans|Raleway&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css" />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
+    />
     <link rel="stylesheet" href="css/animate.css" />
     <link rel="stylesheet" href="css/style.css" />
     <link
@@ -66,18 +69,13 @@
                 <li><a href="#" data-nav-section="about">About</a></li>
                 <li><a href="#" data-nav-section="skills">Skills</a></li>
                 <li>
+                  <a href="#" data-nav-section="experience">Work Experience</a>
+                </li>
+                <li><a href="#" data-nav-section="education">Education</a></li>
+                <li>
                   <a href="#" data-nav-section="repos">Top Repositories</a>
                 </li>
                 <li><a href="#" data-nav-section="blogs">Blogs</a></li>
-                <li>
-                  <a href="#" data-nav-section="experience">Work Experience</a>
-                </li>
-                <li>
-                  <a href="#" data-nav-section="trekking"
-                    >Trekking Adventures</a
-                  >
-                </li>
-                <li><a href="#" data-nav-section="education">Education</a></li>
                 <li><a href="#" data-nav-section="contact">Contact</a></li>
               </ul>
             </div>
@@ -223,41 +221,12 @@
                   <ul
                     id="blogs"
                     class="submenu"
-                    style="
-                      display: flex;
-                      flex-direction: column;
-                      gap: 20px;
-                    "
+                    style="display: flex; flex-direction: column; gap: 20px"
                   ></ul>
                 </li>
               </ul>
             </div>
           </section>
-          <section class="colorlib-skills" data-section="trekking">
-            <div class="colorlib-narrow-content">
-              <div class="row row-bottom-padded-sm animate-box" data-animate-effect="fadeInLeft">
-                <div class="about-desc text-center">
-                  <h1>
-                    Trekking & Himalayan Pass <span style="color: #f9bf3f">Adventures</span>
-                  </h1>
-                  <p>
-                    I’m passionate about exploring the great outdoors and have completed several
-                    memorable treks across the Himalayas.
-                  </p>
-                </div>
-                <div class="row">
-                  <div class="col-md-6">
-                    <h3 class="section-subtitle">Treks Completed</h3>
-                    <div id="trekking" class="adventure-list"></div>
-                  </div>
-                  <div class="col-md-6">
-                    <h3 class="section-subtitle">Mountain Passes Explored</h3>
-                    <div id="passes" class="adventure-list"></div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </section>          
           <section class="colorlib-contact" data-section="contact">
             <div class="colorlib-narrow-content">
               <div
