+++ pages/adventure.html
@@ -1,0 +1,77 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <title></title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="description" content="Portfolio webpage of Vinay Somawat" />
+    <meta
+      name="keywords"
+      content="portfolio,website,personal website,github,projects,computer science and engineering"
+    />
+    <meta name="author" content="Vinay Somawat" />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+    />
+    <link rel="icon" type="image/png" href="../logo.png" />
+    <link
+      href="https://fonts.googleapis.com/css?family=Montserrat|Open+Sans|Raleway&display=swap"
+      rel="stylesheet"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
+    />
+    <link rel="stylesheet" href="../css/animate.css" />
+    <link rel="stylesheet" href="../css/style.css" />
+    <link
+      rel="stylesheet"
+      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css"
+    />
+
+    <script src="../js/modernizr-2.6.2.min.js"></script>
+    <script type="module" src="../index.js"></script>
+    <script type="module" src="../js/main.js"></script>
+    <script type="module" src="adventure.js"></script>
+  </head>
+  <body style="padding: 0 20px">
+    <div class="home-link-wrapper">
+      <a href="../index.html" class="home-link">
+        <i class="fa-solid fa-arrow-left"></i> Home
+      </a>
+    </div>
+    <section class="colorlib-skills" data-section="trekking">
+      <div class="colorlib-narrow-content">
+        <div
+          class="row row-bottom-padded-sm animate-box"
+          data-animate-effect="fadeInLeft"
+        >
+          <div class="about-desc text-center">
+            <h1>
+              Trekking & Himalayan Pass
+              <span style="color: #f9bf3f">Adventures</span>
+            </h1>
+            <p>
+              I’m passionate about exploring the great outdoors and have
+              completed several memorable treks across the Himalayas.
+            </p>
+          </div>
+          <div class="row">
+            <div class="col-md-6">
+              <h3 class="section-subtitle">Treks Completed</h3>
+              <div id="trekking" class="adventure-list"></div>
+            </div>
+            <div class="col-md-6">
+              <h3 class="section-subtitle">Mountain Passes Explored</h3>
+              <div id="passes" class="adventure-list"></div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+    <script src="../js/jquery.min.js"></script>
+    <script src="../js/jquery.waypoints.min.js"></script>
+  </body>
+</html>
